Sort with WASM MUY BONITO PEPITO

+++ src/particles/mod.rs
@@ -18,7 +18,7 @@ pub struct Canvas {
 pub struct Opts {
     // Particles
     #[serde(rename = "NUMBER_OF_PARTICLES")]
-    pub number_of_particles: u16,
+    pub number_of_particles: usize,
     #[serde(rename = "PARTICLE_MAX_SIZE")]
     pub particle_max_size: f32,
     #[serde(rename = "PARTICLE_MIN_SIZE")]
